requirements.txt
# Requirements for Job Scraper

# Core dependencies
crawl4ai>=0.7.8
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Database
sqlite3  # Built-in with Python

# API
python-dotenv>=1.0.0

# Selenium for SPA/Workday pages
selenium>=4.15.0
webdriver-manager>=4.0.1

# Async support
asyncio  # Built-in with Python
aiohttp>=3.9.0